backend/requirements.txt
fastapi[all]
uvicorn[standard]
pydantic
httpx
pytest
black
flake8
pre-commit
python-socketio
fastapi-socketio[asyncio]
python-jose[cryptography]
slowapi
PyJWT
prometheus-fastapi-instrumentator
prometheus-client